requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
email-validator==1.3.0
idna==3.4
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
